requirements.txt
pandas
numpy
yfinance
matplotlib
scipy
python-dotenv
pytest
